+++ dwara-commons/src/main/java/org/ishafoundation/dwaraapi/commandline/local/CommandLineExecuter.java
@@ -8,6 +8,8 @@ import java.util.List;
  */
 public interface CommandLineExecuter {
 
+	public Process createProcess(String command) throws Exception;
+	
 	public Process createProcess(List<String> commandList) throws Exception;
 	
 	public CommandLineExecutionResponse executeCommand(String command) throws Exception;
@@ -17,6 +19,10 @@ public interface CommandLineExecuter {
 	public CommandLineExecutionResponse executeCommand(List<String> commandList) throws Exception;
 	
 	public CommandLineExecutionResponse executeCommand(List<String> commandList, boolean extractLastLineAsFailureReason) throws Exception;
+	
+	public CommandLineExecutionResponse executeCommand(String command, Process proc) throws Exception;
+	
+	public CommandLineExecutionResponse executeCommand(String command, Process proc, boolean extractLastLineAsFailureReason) throws Exception;
 	
 	public CommandLineExecutionResponse executeCommand(List<String> commandList, Process proc) throws Exception;