+++ dwara-videopub/src/main/java/org/ishafoundation/videopub/transcoding/photo/Photo_LowResolution_Transcoding_TaskExecutor.java
@@ -1,4 +1,4 @@
-package org.ishafoundation.videopub.transcoding.image;
+package org.ishafoundation.videopub.transcoding.photo;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -18,11 +18,11 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
-@Component("image-proxy-gen")
+@Component("photo-proxy-gen")
 @Primary
 @Profile({ "!dev & !stage" })
-public class Image_LowResolution_Transcoding_TaskExecutor extends MediaTask implements IProcessingTask{
-    private static final Logger logger = LoggerFactory.getLogger(Image_LowResolution_Transcoding_TaskExecutor.class);
+public class Photo_LowResolution_Transcoding_TaskExecutor extends MediaTask implements IProcessingTask{
+    private static final Logger logger = LoggerFactory.getLogger(Photo_LowResolution_Transcoding_TaskExecutor.class);
     
 	@Override
 	public ProcessingtaskResponse execute(ProcessContext processContext) throws Exception {
@@ -51,13 +51,17 @@ public class Image_LowResolution_Transcoding_TaskExecutor extends MediaTask impl
 		FileUtils.forceMkdir(new File(destinationDirPath));
 	
 		String fileName = FilenameUtils.getBaseName(sourceFilePathname);
-		String thumbnailTargetLocation = destinationDirPath + File.separator + fileName + "_t" + ".jpg";
+		String thumbnailTargetLocation = destinationDirPath + File.separator + fileName + ".thm";
 		String proxyTargetLocation = destinationDirPath + File.separator + fileName + "_p" + ".jpg";
 	
 		long conversionStartTime = System.currentTimeMillis();
 		
-		List<String> proxyCommandParamsList = getProxyCommand(sourceFilePathname, thumbnailTargetLocation, proxyTargetLocation);
-		CommandLineExecutionResponse proxyCommandLineExecutionResponse = commandLineExecuter.executeCommand(proxyCommandParamsList);
+		String proxyCommand = getProxyCommand(sourceFilePathname, thumbnailTargetLocation, proxyTargetLocation);
+		Process proc = commandLineExecuter.createProcess(proxyCommand);
+		CommandLineExecutionResponse proxyCommandLineExecutionResponse = commandLineExecuter.executeCommand(proxyCommand, proc);
+		
+//		List<String> proxyCommandList = getProxyCommandAsList(sourceFilePathname, destinationDirPath, thumbnailTargetLocation, proxyTargetLocation);
+//		CommandLineExecutionResponse proxyCommandLineExecutionResponse = commandLineExecuter.executeCommand(proxyCommandList);
 		
 		long conversionEndTime = System.currentTimeMillis();
 		if(proxyCommandLineExecutionResponse.isComplete()) {
@@ -70,7 +74,11 @@ public class Image_LowResolution_Transcoding_TaskExecutor extends MediaTask impl
 		File xmpSidecarFile = logicalFile.getSidecarFile("xmp");
 		if(xmpSidecarFile != null) {
 			FileUtils.copyFile(xmpSidecarFile, new File(destinationDirPath + File.separator + fileName + ".xmp"));  
+		}else {
+			List<String> extractXmpCommandParamsList = extractXmpCommand(sourceFilePathname, destinationDirPath);
+			CommandLineExecutionResponse extractXmpCommandLineExecutionResponse = commandLineExecuter.executeCommand(extractXmpCommandParamsList);
 		}
+		
 		
 		// TODO : better this...
 		ProcessingtaskResponse processingtaskResponse = new ProcessingtaskResponse();
@@ -84,8 +92,17 @@ public class Image_LowResolution_Transcoding_TaskExecutor extends MediaTask impl
 	}
 
 	// convert 20190716_VVD_0206.NEF \( +clone -resize 192 -quality 50 -write 20190716_VVD_0206-s33.jpg +delete \) -resize 1536 -quality 85 20190716_VVD_0206-s32.jpg
-	private List<String> getProxyCommand(String sourceFilePathname, String thumbnailTargetLocation, String proxyTargetLocation) {
+	private String getProxyCommand(String sourceFilePathname, String thumbnailTargetLocation, String proxyTargetLocation) {
+		return "sh -c 'convert \"" + sourceFilePathname + "\" \\( +clone -resize 192 -quality 50 -write \"JPEG:" + thumbnailTargetLocation + "\" +delete \\) -resize 1536 -quality 85 \"" + proxyTargetLocation + "\"'";
+	}
+
+	private List<String> getProxyCommandAsList(String sourceFilePathname, String destinationDirPath, String thumbnailTargetLocation, String proxyTargetLocation) {
 		List<String> thumbnailGenerationCommandParamsList = new ArrayList<String>();
+		thumbnailGenerationCommandParamsList.add("sh");
+		thumbnailGenerationCommandParamsList.add("-c");
+		thumbnailGenerationCommandParamsList.add("'cd");
+		thumbnailGenerationCommandParamsList.add(destinationDirPath);
+		thumbnailGenerationCommandParamsList.add(";");
 		thumbnailGenerationCommandParamsList.add("convert");
 		thumbnailGenerationCommandParamsList.add(sourceFilePathname);
 		thumbnailGenerationCommandParamsList.add("\\(");
@@ -95,16 +112,29 @@ public class Image_LowResolution_Transcoding_TaskExecutor extends MediaTask impl
 		thumbnailGenerationCommandParamsList.add("-quality");
 		thumbnailGenerationCommandParamsList.add("50");
 		thumbnailGenerationCommandParamsList.add("-write");
-		thumbnailGenerationCommandParamsList.add(thumbnailTargetLocation);
+		thumbnailGenerationCommandParamsList.add("JPEG:" + thumbnailTargetLocation);
 		thumbnailGenerationCommandParamsList.add("+delete");
 		thumbnailGenerationCommandParamsList.add("\\)");
 		thumbnailGenerationCommandParamsList.add("-resize");
 		thumbnailGenerationCommandParamsList.add("1536");
 		thumbnailGenerationCommandParamsList.add("-quality");
 		thumbnailGenerationCommandParamsList.add("85");
-		thumbnailGenerationCommandParamsList.add(proxyTargetLocation);
+		thumbnailGenerationCommandParamsList.add(proxyTargetLocation + "'");
 		
 		return thumbnailGenerationCommandParamsList;
+	}
+	
+	// exiv2 -eX ex 20190207_VVD_0101to0107-mp-e-ot1.tif
+	private List<String> extractXmpCommand(String sourceFilePathname, String destinationDirPath) {
+		List<String> xmpCommandParamsList = new ArrayList<String>();
+		xmpCommandParamsList.add("exiv2");
+		xmpCommandParamsList.add("-eX");
+		xmpCommandParamsList.add("-l");
+		xmpCommandParamsList.add(destinationDirPath);
+		xmpCommandParamsList.add("ex");
+		xmpCommandParamsList.add(sourceFilePathname);
+		
+		return xmpCommandParamsList;
 	}
 
 }